requirements.txt
Flask
discord.py
hypixel
requests
gevent
datetime
bs4
aiohttp
nbt
base64